syscalls_rust::syscalls
Module arch64

This module supports all the 64 bit arch syscalls

Functions

accept⚠

bind⚠

brk⚠

brk() and sbrk() change the location of the program break, which defines the end of the process’s data segment.

chdir⚠

chmod⚠

chown⚠

close⚠

close() closes a file descriptor, so that it no longer refers to any file and may be reused.

connect⚠

creat⚠

execve⚠

exit⚠

exit() terminates the calling process “immediately”.

fchdir⚠

fchmod⚠

fcntl⚠

fdatasync⚠

flock⚠

fork⚠

fork() creates a new process by duplicating the calling process. The new process is referred to as the child process. The calling process is referred to as the parent process.

fsync⚠

ftruncate⚠

getcwd⚠

getegid⚠

geteuid⚠

getgid⚠

getpgrp⚠

getpid⚠

getpid() returns the process ID (PID) of the calling process. (This is often used by routines that generate unique temporary filenames.)

getppid⚠

getuid⚠

kill⚠

The kill() system call can be used to send any signal to any process group or process.

link⚠

listen⚠

lseek⚠

mkdir⚠

mmap⚠

mount⚠

mprotect⚠

munmap⚠

newfstat⚠

newlstat⚠

newstat⚠

open⚠

The open() system call opens the file specified by pathname. If the specified file does not exist, it may optionally (if O_CREAT is specified in flags) be created by open().

poll⚠

read⚠

read() attempts to read up to count bytes from file descriptor fd into the buffer starting at buf.

readlink⚠

rename⚠

rmdir⚠

setgid⚠

setpgid⚠

setsid⚠

setuid⚠

socket⚠

swapoff⚠

swapon⚠

symlink⚠

syslog⚠

truncate⚠

umask⚠

umount⚠

unlink⚠

vfork⚠

vfork - create a child process and block parent

write⚠

write() writes up to count bytes from the buffer starting at buf to the file referred to by the file descriptor fd.