syscalls_rust::syscalls::arch64
Function close

pub unsafe extern "system" fn close(fd: c_uint) -> c_int

close() closes a file descriptor, so that it no longer refers to any file and may be reused.

RETURN VALUE

close() returns zero on success. On error, -1 is returned, and errno is set to indicate the error.

ERRORS

EBADF(9), EINTR(4), EIO(5), ENOSPC(28), etc.

Link

Read the docs here