syscalls_rust::syscalls::arch64
Function write

pub unsafe extern "system" fn write(
    fd: c_uint,
    buf: *const c_char,
    count: size_t,
) -> ssize_t

write() writes up to count bytes from the buffer starting at buf to the file referred to by the file descriptor fd.

RETURN VALUE

On success, the number of bytes written is returned. On error, -1 is returned, and errno is set to indicate the error.

ERRORS

EAGAIN(35), EBADF(9), EDESTADDRREQ(39), EDQUOT(69), EFAULT(14), EFBIG(27),
EINTR(4), EINVAL(22), EIO(5), ENOSPC(28), EPERM(1), EPIPE(32), etc.

Link

Read the docs here